re-scope and re-organize CLI code
Woli::Cli class prevented from using Woli::Cli as a module name
for namespacing all CLI-related code. So Woli::Cli was renamed
to Woli::Cli::Runner and Woli::Cli is now a module.

Parts of the Runner used to be in files named like
'woli/cli/edit.rb', now they are named like 'woli/cli/runner_edit.rb'
to make it more apparent they are parts of the Runner. (They cannot
be defined as modules and included into the Runner, because they
need to access Thor's singleton class methods.)

DateParser was moved into the Woli::Cli module, as it is only supposed
to be used when parsing user's commands.